+++ next.config.js
@@ -15,6 +15,11 @@ const nextConfig = {
   async redirects() {
     return [
       {
+        source: '/',
+        destination: '/swap',
+        permanent: true,
+      },
+      {
         source: '/(B|b)(E|e)(D|d)',
         destination: '/swap',
         permanent: true,
