+++ src/app/products/components/loading-skeleton.tsx
