+++ src/app/products/components/logo-full.tsx
