+++ src/app/products/components/product-row-item/mobile-row.tsx
