+++ src/app/products/utils/formatters.ts
