+++ src/constants/paths.ts
@@ -1,0 +1,5 @@
+export enum Path {
+  HOMEPAGE = 'https://indexcoop.com/',
+  PRODUCTS = '/products',
+  TRADE = '/swap',
+}
